@@ -4,10 +4,11 @@
   "description": "Rest API for Mailpy",
   "main": "src/index.js",
   "scripts": {
-    "start": "node src/index.js",
+    "build": "npm run clean && mkdir -p build && babel src -s -D -d build",
+    "clean": "rm -rf build/",
     "dev": "nodemon --inspect  --exec babel-node src/index.js",
-    "test": "jest",
-    "build": "npm run clean && mkdir -p dist && babel src -s -D -d dist"
+    "start": "node build/index.js",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -55,7 +56,7 @@
         "@babel/env",
         {
           "targets": {
-            "node": "12.19.0"
+            "node": "15.3.0"
           }
         }
       ]
